web_app/requirements.txt
django
psycopg2-binary
requests
django-filter